change score from int to double

+++ src/main/java/dwsc/proyecto/insertmovie/domain/Movie.java
@@ -9,7 +9,7 @@ public class Movie {
 	private String id;
 	private String title;
 	private int year;
-	private int scoreAverage;
+	private double scoreAverage;
 	private String description;
 	private String url;
 
@@ -37,11 +37,11 @@ public class Movie {
 		this.year = year;
 	}
 
-	public int getAverageScore() {
+	public double getAverageScore() {
 		return scoreAverage;
 	}
 
-	public void setAverageScore(int averageScore) {
+	public void setAverageScore(double averageScore) {
 		this.scoreAverage = averageScore;
 	}
